# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winget Repo Manager API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface WingetPackage {
  packageId: string;
  name: string;
  publisher: string;
  version: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  license?: string | null;
  /** @nullable */
  homepage?: string | null;
}

export interface LocalPackage {
  id: number;
  packageId: string;
  name: string;
  publisher: string;
  version: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  license?: string | null;
  /** @nullable */
  homepage?: string | null;
  addedAt: string;
}

export interface AddPackageBody {
  packageId: string;
  name: string;
  publisher: string;
  version: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  license?: string | null;
  /** @nullable */
  homepage?: string | null;
}

export interface PackageStats {
  total: number;
  publishers: number;
  recentlyAdded: number;
}

export type SearchWingetParams = {
  q: string;
  limit?: number;
};