# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winget Repo Manager API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Search official Winget community repo
 */
export const searchWingetQueryLimitDefault = 30;

export const SearchWingetQueryParams = zod.object({
  q: zod.coerce.string(),
  limit: zod.coerce.number().default(searchWingetQueryLimitDefault),
});

export const SearchWingetResponseItem = zod.object({
  packageId: zod.string(),
  name: zod.string(),
  publisher: zod.string(),
  version: zod.string(),
  description: zod.string().nullish(),
  license: zod.string().nullish(),
  homepage: zod.string().nullish(),
});
export const SearchWingetResponse = zod.array(SearchWingetResponseItem);

/**
 * @summary List packages in the local repo
 */
export const ListPackagesResponseItem = zod.object({
  id: zod.number(),
  packageId: zod.string(),
  name: zod.string(),
  publisher: zod.string(),
  version: zod.string(),
  description: zod.string().nullish(),
  license: zod.string().nullish(),
  homepage: zod.string().nullish(),
  addedAt: zod.coerce.date(),
});
export const ListPackagesResponse = zod.array(ListPackagesResponseItem);

/**
 * @summary Add a package to the local repo
 */
export const AddPackageBody = zod.object({
  packageId: zod.string(),
  name: zod.string(),
  publisher: zod.string(),
  version: zod.string(),
  description: zod.string().nullish(),
  license: zod.string().nullish(),
  homepage: zod.string().nullish(),
});

/**
 * @summary Get a single package
 */
export const GetPackageParams = zod.object({
  id: zod.coerce.number(),
});

export const GetPackageResponse = zod.object({
  id: zod.number(),
  packageId: zod.string(),
  name: zod.string(),
  publisher: zod.string(),
  version: zod.string(),
  description: zod.string().nullish(),
  license: zod.string().nullish(),
  homepage: zod.string().nullish(),
  addedAt: zod.coerce.date(),
});

/**
 * @summary Remove a package from the local repo
 */
export const RemovePackageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get package statistics summary
 */
export const GetPackageStatsResponse = zod.object({
  total: zod.number(),
  publishers: zod.number(),
  recentlyAdded: zod.number(),
});