# lib/api-zod/src/generated/types/localPackage.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winget Repo Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface LocalPackage {
  id: number;
  packageId: string;
  name: string;
  publisher: string;
  version: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  license?: string | null;
  /** @nullable */
  homepage?: string | null;
  addedAt: Date;
}