# lib/api-zod/src/generated/types/packageStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Winget Repo Manager API
 * OpenAPI spec version: 0.1.0
 */

export interface PackageStats {
  total: number;
  publishers: number;
  recentlyAdded: number;
}